Classifiers/requirements-e.txt
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.4
chainer==7.7.0
chardet==4.0.0
cupy-cuda110==8.2.0
dataclasses==0.8
decorator==4.4.2
fastrlock==0.5
filelock==3.0.12
future==0.18.2
idna==2.10
joblib==1.0.0
jsonpatch==1.28
jsonpointer==2.0
librosa==0.7.2
llvmlite==0.31.0
numba==0.48.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1604945999799/work
pandas==1.0.3
Pillow==8.0.1
protobuf==3.14.0
pycparser==2.20
python-dateutil==2.8.1
pytorch-ignite==0.3.0
pytz==2020.4
pyzmq==20.0.0
requests==2.25.1
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.4.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
SoundFile==0.10.3.post1
termcolor==1.1.0
torch==1.7.1
torchaudio==0.7.2
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.8.2
tornado==6.1
tqdm==4.43.0
typing-extensions==3.7.4.3
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0
